+++ waspc/src/Wasp/Analyzer/Prisma.hs
@@ -7,6 +7,7 @@ where
 import Wasp.Analyzer.Parser as Parser
 import qualified Wasp.Psl.Ast.Model as Psl.Model
 import qualified Wasp.Psl.Ast.Schema as Psl.Schema
+import qualified Wasp.Psl.Ast.WithCtx as Psl.WithCtx
 import qualified Wasp.Psl.Generator.Model as Psl.Model.Generator
 
 injectEntitiesFromPrismaSchema :: Psl.Schema.Schema -> Parser.AST -> Either a Parser.AST
@@ -36,8 +37,6 @@ makeEntityStmt (name, body) = wrapWithCtx $ Parser.Decl "entity" name $ wrapWith
 -- | Generates Prisma models source code so that it can be injected into Wasp AST.
 generatePrismaModelSources :: Psl.Schema.Schema -> [(ModelName, ModelBody)]
 generatePrismaModelSources schema =
-  [ ( name,
-      Psl.Model.Generator.generateModelBody body
-    )
-    | (Psl.Model.Model name body) <- Psl.Schema.getModels schema
+  [ (name, Psl.Model.Generator.generateModelBody body)
+    | (Psl.Model.Model name body) <- Psl.WithCtx.getNode <$> Psl.Schema.getModels schema
   ]