+++ waspc/src/Wasp/Psl/Generator/Model.hs
@@ -6,9 +6,10 @@ where
 import qualified Wasp.Psl.Ast.Model as Psl.Model
 import Wasp.Psl.Generator.Attribute (generateAttribute)
 import Wasp.Psl.Generator.Common (PslSource)
+import Wasp.Psl.Generator.WithCtx (generateWithCtx)
 
 generateModelBody :: Psl.Model.Body -> PslSource
-generateModelBody (Psl.Model.Body elements) = unlines $ map (("  " ++) . generateModelElement) elements
+generateModelBody (Psl.Model.Body elements) = unlines $ generateWithCtx generateModelElement <$> elements
 
 generateModelElement :: Psl.Model.Element -> PslSource
 generateModelElement (Psl.Model.ElementField field) =