+++ waspc/src/Wasp/Psl/Parser/Model.hs
@@ -22,6 +22,7 @@ import Wasp.Psl.Parser.Common
     symbol,
     whiteSpace,
   )
+import Wasp.Psl.Parser.WithCtx (withCtx)
 
 -- | This is used to parse the body of the PSL tags in the Wasp file.
 -- NOTE: We need to consume the leading whitespace specifically here, because we use the `body`
@@ -47,7 +48,7 @@ model = do
 -- which is everything besides model keyword, name and braces:
 --   `model User { <body> }`.
 body :: Parser Psl.Model.Body
-body = Psl.Model.Body <$> some element
+body = Psl.Model.Body <$> some (withCtx element)
 
 element :: Parser Psl.Model.Element
 element =