+++ waspc/test/Psl/Parser/EnumTest.hs
@@ -7,6 +7,7 @@ import qualified Text.Megaparsec as Megaparsec
 import qualified Wasp.Psl.Ast.Argument as Psl.Argument
 import qualified Wasp.Psl.Ast.Attribute as Psl.Attribute
 import qualified Wasp.Psl.Ast.Enum as Psl.Enum
+import qualified Wasp.Psl.Ast.WithCtx as Psl.WithCtx
 import qualified Wasp.Psl.Parser.Enum as Psl.Parser
 
 spec_parsePslEnum :: Spec
@@ -26,18 +27,19 @@ spec_parsePslEnum = do
           expectedAst =
             Psl.Enum.Enum
               "Role"
-              [ Psl.Enum.ElementValue "USER" [],
-                Psl.Enum.ElementValue
-                  "ADMIN"
-                  [ Psl.Attribute.Attribute
-                      "map"
-                      [Psl.Argument.ArgUnnamed $ Psl.Argument.StringExpr "ADMIN_MAPPING"]
-                  ],
-                Psl.Enum.ElementBlockAttribute $
-                  Psl.Attribute.Attribute
-                    "map"
-                    [Psl.Argument.ArgUnnamed $ Psl.Argument.StringExpr "enum_name"]
-              ]
+              $ Psl.WithCtx.empty
+                <$> [ Psl.Enum.ElementValue "USER" [],
+                      Psl.Enum.ElementValue
+                        "ADMIN"
+                        [ Psl.Attribute.Attribute
+                            "map"
+                            [Psl.Argument.ArgUnnamed $ Psl.Argument.StringExpr "ADMIN_MAPPING"]
+                        ],
+                      Psl.Enum.ElementBlockAttribute $
+                        Psl.Attribute.Attribute
+                          "map"
+                          [Psl.Argument.ArgUnnamed $ Psl.Argument.StringExpr "enum_name"]
+                    ]
       Megaparsec.parse Psl.Parser.enum "" source `shouldBe` Right expectedAst
 
     it "Commented out fields" $ do
@@ -54,10 +56,11 @@ spec_parsePslEnum = do
           expectedAst =
             Psl.Enum.Enum
               "Role"
-              [ Psl.Enum.ElementValue "USER" [],
-                Psl.Enum.ElementBlockAttribute $
-                  Psl.Attribute.Attribute
-                    "map"
-                    [Psl.Argument.ArgUnnamed $ Psl.Argument.StringExpr "enum_name"]
-              ]
+              $ Psl.WithCtx.empty
+                <$> [ Psl.Enum.ElementValue "USER" [],
+                      Psl.Enum.ElementBlockAttribute $
+                        Psl.Attribute.Attribute
+                          "map"
+                          [Psl.Argument.ArgUnnamed $ Psl.Argument.StringExpr "enum_name"]
+                    ]
       Megaparsec.parse Psl.Parser.enum "" source `shouldBe` Right expectedAst