+++ waspc/test/Psl/Parser/ModelTest.hs
@@ -9,6 +9,7 @@ import qualified Text.Megaparsec as Megaparsec
 import qualified Wasp.Psl.Ast.Argument as Psl.Argument
 import qualified Wasp.Psl.Ast.Attribute as Psl.Attribute
 import qualified Wasp.Psl.Ast.Model as Psl.Model
+import qualified Wasp.Psl.Ast.WithCtx as Psl.WithCtx
 import Wasp.Psl.Parser.Model (model, parseBody)
 
 spec_parsePslModel :: Spec
@@ -52,26 +53,27 @@ spec_parsePslModel = do
         modelAst =
           Psl.Model.Model
             "User"
-            ( Psl.Model.Body
-                [ Psl.Model.ElementField
-                    ( Psl.Model.Field
-                        "id"
-                        Psl.Model.Int
-                        []
-                        [ Psl.Attribute.Attribute "id" [],
-                          Psl.Attribute.Attribute
-                            "default"
-                            [ Psl.Argument.ArgNamed "value" (Psl.Argument.FuncExpr "autoincrement" [])
-                            ]
-                        ]
-                    ),
-                  Psl.Model.ElementField
-                    ( Psl.Model.Field "internets" (Psl.Model.UserType "Internet") [Psl.Model.List] []
-                    ),
-                  Psl.Model.ElementField
-                    ( Psl.Model.Field "strings" (Psl.Model.UserType "Strings") [Psl.Model.List] []
-                    )
-                ]
+            ( Psl.Model.Body $
+                Psl.WithCtx.empty
+                  <$> [ Psl.Model.ElementField
+                          ( Psl.Model.Field
+                              "id"
+                              Psl.Model.Int
+                              []
+                              [ Psl.Attribute.Attribute "id" [],
+                                Psl.Attribute.Attribute
+                                  "default"
+                                  [ Psl.Argument.ArgNamed "value" (Psl.Argument.FuncExpr "autoincrement" [])
+                                  ]
+                              ]
+                          ),
+                        Psl.Model.ElementField
+                          ( Psl.Model.Field "internets" (Psl.Model.UserType "Internet") [Psl.Model.List] []
+                          ),
+                        Psl.Model.ElementField
+                          ( Psl.Model.Field "strings" (Psl.Model.UserType "Strings") [Psl.Model.List] []
+                          )
+                      ]
             )
 
     it "Model is correctly parsed" $ do
@@ -93,14 +95,15 @@ spec_parsePslModel = do
         modelAst =
           Psl.Model.Model
             "User"
-            ( Psl.Model.Body
-                [ Psl.Model.ElementField (Psl.Model.Field "model" Psl.Model.Int [] []),
-                  Psl.Model.ElementField (Psl.Model.Field "type" Psl.Model.String [] []),
-                  Psl.Model.ElementField (Psl.Model.Field "view" Psl.Model.Boolean [] []),
-                  Psl.Model.ElementField (Psl.Model.Field "enum" Psl.Model.Float [] []),
-                  Psl.Model.ElementField (Psl.Model.Field "generator" Psl.Model.Decimal [] []),
-                  Psl.Model.ElementField (Psl.Model.Field "datasource" Psl.Model.String [] [])
-                ]
+            ( Psl.Model.Body $
+                Psl.WithCtx.empty
+                  <$> [ Psl.Model.ElementField (Psl.Model.Field "model" Psl.Model.Int [] []),
+                        Psl.Model.ElementField (Psl.Model.Field "type" Psl.Model.String [] []),
+                        Psl.Model.ElementField (Psl.Model.Field "view" Psl.Model.Boolean [] []),
+                        Psl.Model.ElementField (Psl.Model.Field "enum" Psl.Model.Float [] []),
+                        Psl.Model.ElementField (Psl.Model.Field "generator" Psl.Model.Decimal [] []),
+                        Psl.Model.ElementField (Psl.Model.Field "datasource" Psl.Model.String [] [])
+                      ]
             )
 
     it "Model is correctly parsed" $ do